refactor(scout-for-lol): isolate Dare settlement reporting
Why

Keep the Dare v2 settlement coordinator within the repository's module-size boundary as compiled proof metadata is layered onto the stack.

What

Move batch inspection and settlement failure logging plus Sentry reporting into a focused helper module without changing recovery behavior.

Verification

- bunx eslint src/betting/dare-settle-v2.ts src/betting/dare-settle-report-v2.ts
- bunx lefthook run pre-commit
